Use /run instead of /tmp to fix insecure tempfiles
Introduces a new required setting "uid" which is used to construct the
path to /run/user/<uid>/nsh/job-*.{root,stderr}. This is more secure
than using files created in /tmp.

Potential alternative discussion:

1) Using a persistent directory in a place like /var is no good because
it's possible for an attacker to create such a directory before NSH is
ever run on a particular system. Because Slurm is what will end up
creating the directory, there is no way to ensure that the directory
does not already exist or has the right ownership/permissions.

2) Using mktemp / mkstemp is also not viable because the path to the
stderr file must be computed before job submission, which means before
it is known what host the job will run on.

Resolves #8

Signed-off-by: Lisanna Dettwyler <lisanna.dettwyler@gmail.com>